Include RC module to pull out configuration
Right now on a fresh pull of master I’m unable to simply start the app
by running npm start because it’s expecting a configuration file which
doesn’t exist in source control. This file is not under source control
because it contains passwords. This change updates the readme w/
instructions and delegates that responsibility a node module plugin
called RC. The power the plugin provides is it has a hierarchy of
places it looks it’s looks for a configuration file providing a
mechanism to override on the server.